+++ src/routes/librarian-routes/fine-payment.ts
@@ -1,0 +1,10 @@
+import { Router } from "express";
+import { getFinePaymentHandler } from "../../controller/librarian/fine-payment/fine-payment.controller";
+import { validate } from "../../middleware/validate";
+import { paginationSchema } from "../../validations/pagination.schema";
+
+const router = Router();
+
+router.get("/", validate(paginationSchema), getFinePaymentHandler);
+
+export default router;